sessions user id

+++ public/js/delete-post.js
@@ -4,16 +4,23 @@ async function deleteFormHandler(event) {
     const id = window.location.toString().split('/')[
       window.location.toString().split('/').length - 1
     ];
-    const response = await fetch(`/api/posts/${id}`, {
-      method: 'DELETE'
-    });
+    if (1 === 1) {
+      
+    
+      const response = await fetch(`/api/foods/${id}`, {
+        method: 'DELETE'
+      });
+
+      
+    }
   
     if (response.ok) {
       document.location.replace('/dashboard/');
     } else {
       alert(response.statusText);
     }
+    //alert('this is working');
   }
   
-  document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
+  document.getElementById('delete-btn').addEventListener('click', deleteFormHandler);
